Fixup rate limiter with redis < 7.0
This was missed when implementing the initial rate limiter, we cannot use NX on the EXPIRE operation, because that's a 7.0+ operator only. Sourcegraph ships with redis 5 by default, so that will fail on dotcom.

Keegan and I brainstormed this approach together and concluded that we don't need the MULTI statement and do all in one go, so we can simply read the TTL and set expiry when it's not yet set.