Embeddings: don't reload config in middle of job (#51114)
Before, the embeddings client live-reloaded the config whenever it
changed. This wasn't safe, since parts of the config (number of
dimensions, model, etc.) could change in the middle of a method. The
config could also change in the middle of a job, leading to an
inconsistent index state.

This commit improves things a bit by removing the live-reloading. Now
the config stays the same for the duration of each request.