Cody settings trailing slash fixed (#51163)
This PR is for issue #51155. While adding the codebase name with a
trailing "/" it was showing as an error. This PR fixed this bug, now it
accepts codebase name either with "/" or not with it.


## Test plan
All the unit and integration tests have been passed.
Tested the behaviour manually it works fine.