first draft
This adds a test that measures recall for a static set of queries and a
static embeddings index. It is ignored by default, but it can be enabled
by calling `go test` with `-tags sourcegraphCI`.

The motivation is to calculate recall on CI whenever the embeddings
service changes.

For CI we have to provide the embeddings of the queries (we don't want
to call openAI for every test run) and the embeddings of the repo.

The repo embeddings can be generated by triggering an embeddings job in
a local Sourcegraph instance. The query embeddings can be generated with
the help of a little helper script I added in this PR.