Adjust feature flag logic for cody on dotcom
This PR changes to no access to unauthenticated users. We want to hide cody on dotcom. I've also added tests to verify things behave alright.
The feature flag was currently not correct. The backend made a different determination than the frontend did, because the frontend simply looked a the feature flag, while the backend looks at the `ExperimentalFeatures.CodyRestrictUsersFeatureFlag` field as well.
This aligns the two for now by exposing codyEnabled on the jscontext and using the same logic in the backend for that.
This is likely not the end state we want here, but it fixes an immediate problem where unauthed users would be able to see Cody but then aren't able to use it.