App: Add sourcegraph:// url scheme handler (#51193)
Add the ability for App to register a handler for scheme urls that start
with `sourcegraph://`

The url can then be used to open a specific page in the App (also known
as deep linking).

If the app isn't already running, opening a `sourcegraph://` url will
launch it. The url's path will be opened inside of the app's webview.

The implementation relies on an entry in `Info.plist` on mac, and for
that reason it doesn't work using `sg start dev`; it works when doing a
release build.

Tested on Mac so far. There may be tweaks needed for non-mac for initial
launch, I'll test on Linux when I get a chance and that may be a
followup PR.

Closes #51270