dev: add a command to understand search job plans
I was trying to understand exactly how a query I write is transalated
into a zoekt query. So this adds a command which allows me to easily try
stuff out and then see the s-expression output of the jobs plan.

  $ go run ./dev/internal/cmd/search-plan -dotcom 'content:"hello\\nworld" patternType:literal'
  plan ["patterntype:literal" "hello\\nworld"]
  query (and "patterntype:literal" "hello\\nworld")
  (LOG
    (ALERT
      (query . (and "patterntype:literal" "hello\\nworld"))
      (originalQuery . content:"hello\\nworld" patternType:literal)
      (patternType . literal)
      (TIMEOUT
        (timeout . 20s)
        (LIMIT
          (limit . 500)
          (PARALLEL
            (ZOEKTGLOBALTEXTSEARCH
              (query . content_substr:"hello\\nworld")
              (type . text)
              )
            (REPOSCOMPUTEEXCLUDED
              )
            NoopJob)))))

Test Plan: A simple unit test is added to ensure this command doesn't
break.