embeddings: metric for search throughput (#51461)
I was playing around with the benchmark and was interested in the
throughput of the implementations. This is the numbers we seem to be
quoting in text which I assume previously we calculated manually.

Test Plan: go test -bench Search -run '^$'

(cherry picked from commit dfd6a82b15121b60b89e96d008213ae218f19132)