dev: add a command to understand search job plans (#51466)
I was trying to understand exactly how a query I write is transalated
into a zoekt query. So this adds a command which allows me to easily try
stuff out and then see the s-expression output of the jobs plan.

Test Plan: A simple unit test is added to ensure this command doesn't
break.