Embeddings: use a fixed-memory cache (#51384)
Right now, we use an LRU cache with a fixed size of 5 repo embeddings.
This is not very flexible and will perform very poorly when we have a
large number of small repos that cause larger repos to be evicted.

This wraps the LRU cache with a cost-aware cache that allows us to
cap the maximum memory usage of the cache. This allows us to keep a
large number of small repos in memory without forcing large, frequently
used, repos out.