Embeddings: simplify embeddings search index (#51435)
This is just a small simplification to searchRepoEmbeddingIndex. We do
not need to check ahead of time whether there are any embeddings or
whether the limit is zero. We already check in index.SimilaritySearch
for numResults == 0, and to make it more obvious that this does not
change any behavior, I moved the check for zero embeddings to the same
place.

(cherry picked from commit 39a229b403ea370981cad7fdfa9b4aaf3c206a5c)