Embeddings: add repo source to results (#51476)
To support searching multiple repos with embeddings, we need to add the
source repo to the results.

I want to avoid lazily hydrating repo name and revision like we do with
the file content. I prefer structs to be fully constructed on creation.
So to support this, I added a new intermediate `SimilaritySearchResult`
type that only contains the information that is available at the time of
doing the similarity search. We then convert it to the final search
result type `EmbeddingsSearchResult` afterwards, which adds fields like
`RepoName`, `Revision`, and `Content`.

(cherry picked from commit 8499e0956a01fa755097e3785d49ce36949a9b73)