feat: fetch github app data on github app page (#51526)
## Description

Adding a prototype for github app screen, which allows us to do various
actions.

I'm saying a prototype because that's what it is. The code is
intentionally left in this state, since I expect a lot of changes to
happen to this screen next week. Important thing is that most of the
functionality is there.

Last missing thing is tying it up with a new connection, but I decided
to push it to a followup PR, since this one is already too big to my
liking.

## Loom recording

https://www.loom.com/share/165e26cf46f34b00939af79253cf324b

## Screenshot
<img width="922" alt="Screenshot 2023-05-05 at 18 23 54"
src="https://user-images.githubusercontent.com/9974711/236513921-14176811-034e-463d-9053-de19a7cc7c3e.png">


## Test plan

Tested locally. This change is adding to an unreleased feature, so
should be quite safe to merge.