Cody completions: Don't bubble errors (#51580)
This PR does two thinks (albeit it would not need one of them to be
fair):

- Don't bubble errors on inline completions to the user. Completions
trigger very very often which means that any error that happens will
likely repeat a looooot of times. We had this with rate limiting and
with file not found errors and I think it's better for the user to fail
silently, especially in this stage. This also mimics GitHub copilot that
doesn't say a word when it e.g. can't connect to the backend properly.
- Properly unwrap rate limiting errors: Since we don't use the errors
this is not 100% necessary however we know that the response won't parse
to JSON so it still makes sense to handle this case properly. We might
also revisit the first point and then we'll at least have better looking
error message.

## Test plan

Tried it locally, no more errors. Bliss.

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->