[codenav] adding prototypes to  codenav (#50667)
This is a draft PR to address the problem with our current
implementation of "Find Implementations" as described in [this
issue](https://github.com/sourcegraph/sourcegraph/issues/50110).

>Currently, "Find implementations" shows the prototype instead of
implementations. When a symbol A# implements symbol B# then I expect
"Find implementations" on the symbol B# to return A#.

Feedback welcome, but more importantly need to ensure this actually
solves the problem highlighted by TwoSigma.


![animalsExample](https://user-images.githubusercontent.com/6225160/235517415-3f27c37e-5e35-4bed-8082-ee5a0e749311.png)




## Test plan
Manually tested the affected routes and responses to ensure no
regressions. Updated tests as well.

Please review the changes and provide any feedback or suggestions. If
there are any concerns or further adjustments needed, let me know and
I'll address them promptly.
<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->