insights: fix issue where scoped insights would not resolve repos with permissions (#51657)
This fixes a long standing and very mysterious customer issue that has
plagued us for a few months. The underlying problem is that repositories
with permissions will not resolve for insights defined with permissions.

The background here is that this process used to happen in the request
scope of a user, and the requisite search queries to formulate an
insight backfill were derived using the repo names directly. When we
migrated this process to a background routine (due to some insights
being uncreatable in the request timeout) this edge case slipped through
the cracks going from request scope to background scope. Fortunately for
us there was a workaround, which was to switch to the newer repository
query to select the repositories. This worked because the internal
context was set prior to executing the search to do that lookup.

Repo permissions have always been respected at viewer time, and nothing
changes here and was not broken.

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->

## Test plan
I've embedded a test that will verify the scoped backfiller has the
appropriate behavior given internal and non-internal users. I cannot for
the life of me figure out how to set up repo permissions locally, so
will test this manually when it lands on an instance.

Here is an example of the GraphQL response showing validation fails with
an invalid repo
<img width="885" alt="CleanShot 2023-05-09 at 10 48 02@2x"
src="https://github.com/sourcegraph/sourcegraph/assets/5090588/5ecbadf6-4cea-4561-b1cc-f58df575046e">