app: replace nonce with in-memory secret
Previously, the Sourcegraph app relied on a nonce to perform passwordless
authentication. That is, we kept an in-memory secret token, and once the
sign in URL was visited with that secret token, it was invalidated and
you would need the newly generated secret in order to sign in again.

That behavior made sense when the app was merely opening the user's browser
to that URL: they would always have a session and would not need to
authenticate in the future.

When we migrated to Tauri, we kept the same behavior: we would print the
sign in URL to stdout, our Tauri app would see that sign in URl and load
it in the Tauri browser. However, that only works once: we cannot create
more windows after that, as Tauri windows do not share the same session.

We want to create multiple Tauri windows: for the Cody chat system tray
UI, as well as to make app Settings a standalone window - but when we
create those windos and load the sign in URL we got when the server started..
it no longer works. The nonce secret was invalidated / used.

This PR simply switches us from nonces to an in-memory secret. The sign
in URL becomes valid for as long as the process is running. The nonce
behavior was, to my knowledge, just us being more secure - there's no
real harm in us doing the same thing and simply not invalidating it so
that future sign in attempts succeed.

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>