Add additional field resolver for the most recent token
Instead of relying purely on the ordering, this makes it more explicit. The main driver for this is though that we don't want to error out if some of the keys in the list still exist, but the UI depends on that. This is a simple fix for it so that the proxy can list all the valid tokens alright from dotcom.