Refactor: Make introducing directory/repo graphQL API easy (#51637)
Part of #50800

Refactoring of graph QL resolver for ownership - no behavior changes.

This makes it easier to implement similar API for tree resolver
(directory) and commit resolver (repo view), because the same components
that are extracted here can be directly reused. At the end of the day
pagination for connection and signal computation can work the same for
all views.

Includes the following:

- Extract `computeCodeowners` similar to
`computeRecentContributorSignals`,
- Extract `ownershipConnection` that handles ordering and pagination for
both code owners and signals,
- _nit_ Remove `resolvedOwners` from the connection object - it was not
needed.

Caveat: Filed
[#51636](https://github.com/sourcegraph/sourcegraph/issues/51636) to
track ordering improvements. For now the connection ordering is
identifier based and alphabetical. This needs to be improved for better
UX.

## Test plan

Extend the signals test to make sure the refactoring does not break
identity of owner for signals.