embeddings: calculate recall on CI (#51218)
We calculate recall in a Go test whenever the embeddings service
changes. The output from the calculation is available as build artifact
in Buildkite.

The test relies on static embeddings of the Sourcegraph repo and the
queries, both of which are not committed to the repository. Hence `go
test ./...` fails. However `bazel test //...` works. The static
embeddings are stored in a GCP bucket and downloaded by Bazel if
required.

New repo embeddings can be generated by triggering an embeddings job in
a local Sourcegraph instance. The query embeddings can be generated with
the help of a little helper script I have included in this PR.

---------

Co-authored-by: Jean-Hadrien Chabran <jh@chabran.fr>