cody: git-log recipe limits log to active file (#51646)
This seems like a more useful use of git log in a larger codebase.

Test Plan: clicked on the newly added option in the git log recipe and
it summarised changes for the file. Compared to git log output of file
to ensure the response made sense. Then compared it to the repository
git log/etc just to make sure Cody was talking about the file.