frontend: handle http.ErrAbortHandler in HandlerWithErrorReturn
I was investigating a panic in production and the root cause was the
debugproxies reverse proxy failing due to the connection being closed.
In that case reverse proxy panics with ErrAbortHandler. According to the
documentation this is an intentional panic error that indicates to stop
processing and not report the error. As such we do the same in our
top-level error middleware.

Test Plan: I didn't try trigger the code path, instead I am relying on
CI to ensure nothing breaks. Will attach this PR to the error reported
in google cloud. If this error every happens again it will show up as
this PR and we know this was not the solution.