feat: add a message for auth provider config for github apps (#51709)
## Description

GitHub Apps are distributed with OAuth credentials for users to sign in
with. We need to add an auth provider for each github app. Modifying the
site config automatically is not really an option, because:
1. A lot of customers use site config from a file, which would override
any changes that we make
2. Modifying the site config on the backend is really hard, since it's
jsonc and we want to keep any comments. E.g. we cannot just json.Marshal
a struct in Go
3. Even if we could do the above, we only allow 1 auth provider per url.
So we cannot have more than one `https://github.com` auth providers.
Letting the site admin decide what to do in that case seems like the
best bet.

This PR adds a feature that shows to the site admin the configuration
that needs to be added to the site configuration instead. We rely on the
site-admin to do it manually. Given the constraints above, this was the
quickest way to make it functional with the least amount of friction.

## Screenshots

<img width="922" alt="Screenshot 2023-05-10 at 14 50 46"
src="https://github.com/sourcegraph/sourcegraph/assets/9974711/3fd33906-75f9-4b7c-9598-89ed97a733d7">

![Screenshot 2023-05-10 at 14 50
56](https://github.com/sourcegraph/sourcegraph/assets/9974711/5d45d4af-4fd1-45b7-8ab9-e1db5815483d)

## Test plan

Tested locally. Most of the changes are in the UI.