Add handling for 429 Retry-After errors in httpcli
This PR adds some basic handling of the Retry-After for 429 errors. Right now, we retry every 429 error, regardless of if the Retry-After header tells us to wait another hour, where 20 retry attempts would not help at all.
We do that by parsing the header and checking if within the next 3s we would regain access. If so, we keep trying, otherwise, we stop retrying.