Return the output of the sync command in VCSSyncer::Fetch  (#51754)
Administrating Perforce depot syncing in Sourcegraph is made much easier
by being able to examine the output of the `p4-fusion` command. See [PR
51598](https://github.com/sourcegraph/sourcegraph/pull/51598) for
details.

This PR adds support for capturing the output when a background sync
runs. Because it modifies the signature of `VSCSyncer`, it's being
broken out into a separate PR.

## Test plan

existing tests updated to support the returned output

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->