Cody: respect the trace=1 URL param for Cody completions (#51839)
I frequently find myself wanting to view a trace of Cody's response for
ad-hoc measurements, debugging, and exploratory hacking. However,
AFAICT, there is no way right now to get a trace for a Cody query.

This makes Cody in the web UI respect the `trace=1` URL parameter that
is used for search. A URL pointing to the trace ID is the set on the
response header of the `stream` request (just as we do with search
queries).

The traces aren't terribly useful right now, but I plan to follow up by
adding useful information like token sizes, logs whenever we stream a
message, etc.