Propagate tracing information (#51847)
Previously, we were not propagating the parent trace ID across HTTP
requests. This didn't used to be the case -- I'm honestly not sure
when/how this changed, but it's been broken for at least a couple of
months.

This wraps internal requests with the Opentelemetry transport, which
will pass along the tracing information so that we can effectively
connect traces across multiple services together.