Tracing: remove deprecated methods (#51859)
This does the minimum amount of work needed to remove the deprecated
methods on `trace.Trace`. We still rely on the `otlog.Field` type
heavily in the observation package, so to minimize the surface area of
the change, I exported the `OTLogFieldsToOTelAttrs` field so we can
start to push the opentelemetry attribute types upwards in the stack
more incrementally.

## Test plan

It compiles + existing tests.

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->