Tracing: use opentelemetry attrs in observation package (#51911)
This migrates most of the interface of the observation package to use
the opentelemetry attrs rather than the opentracing fields. The only
remaining API surface that uses the opentracing fields is the Args
struct, which is admittedly the most used piece. This PR knocks out the
rest of the low-hanging fruit before tackling the tougher stuff.