Add usage graphs for LLM proxy based on BigQuery (#51863)
This pulls the BigQuery events on the dotcom side and renders a nice
graph. We probably want a configurable window or so, but I'm not super
sure here yet so let's follow up on that. For this to work, we will
probably need some additional terraform magic, and will need to modify
the site-config of dotcom to configure this.