Unify API for code completions and chat
This PR makes the input / output of these two endpoints the same, making them effectively the same thing, with the only difference that another model is chosen for code, and that code completions are non-streaming.
This will be useful long-term as there's less divergence, and help make the implementation of additional upstream providers easier as only one method needs to be implemented in most cases.