Cody compeltions: Disregard whitespace in the completion prefix for the cache (#51920)
This fixes a bug that is currently present in the completions cache
where a reduction of whitespace would cause to additional completion
requests.

To reproduce the bug, imagine this scenario:

```ts
function foo() {
  |
```

With the curser at `|`.

We request completions and cache it with the prefix `function foo() {\n
`. Cool.

However, the unexpected behavior comes when you _remove some whitespace
from the prefix_. So e.g. you Remove one backspace and make it:

```ts
function foo() {
 |
```

With the previous cache, this was resulting in a miss and a new
completion being sent which led to some unexpected behavior, especially
if the IDE inserts whitespace for you and you fix it by pressing
backspace:



https://github.com/sourcegraph/sourcegraph/assets/458591/db9034a0-6545-4e2b-9af0-78bc02cc7863

To fix this, the cache now trims all whitespace _in the last line_
(leaving the total numbers of `\n` unchanged).


## Test plan

- Tested locally by reproducing the above steps
- Added unit tests to the cache

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->