llmproxy: Add OpenAI support (#51952)
This adds support for another upstream model: GPT.

On the LLM Proxy side, this adds a new endpoint that speaks OpenAI,
takes OpenAI configuration, and makes the handling and event logging
mostly generic so that we have more code sharing between anthropic and
openai. On the Sourcegraph side, this adds a new switch in the LLMProxy
client that detects based on the model provided which upstream client to
wrap in the LLM Proxy doer.