Cody: Tweak "connect to App" feature and remove experimental flag (#51949)
Change the App button behavior in the Cody VSCE:

- Show "Connect to Sourcegraph App" always when App is detected as being
installed, without being behind a feature flag. (This detects new
upcoming releases of App only)
- Do not show any App download CTA when it's not installed. We will
re-add the App download CTA when we have a release of App that we're
ready to share more widely.

This removes the "experimental.connectToApp" settings flag in the Cody
extension.

Also, improve the App detector, changing the known App locations to be
per-platform (supporting mac for now).