app: rename 'Sourcegraph App' -> 'Sourcegraph' (#52087)
Naming conventions for now:

* `Sourcegraph App` (capital A) is the product name; when we need to
distinguish the app from e.g. an enterprise version of the product, we
use `Sourcegraph App`.
* `the Sourcegraph app` (lowercase A) is to be used in general
documentation strings or when writing about it in english, in a context
where we do not need to distinguish it from our other product lines.
* `Sourcegraph` is to be used in places where it is obvious it is an
app, e.g. `Sourcegraph.app` instead of `Sourcegraph App.app`, or if we
were to start distributing it in say an app store.

## Test plan

N/A, rename only

---------

Signed-off-by: Stephen Gutekanst <stephen@sourcegraph.com>