embeddings: Move splitter behind codeintel service boundary (#52169)
Should keep behavior the same, but moves the `embeddings/split` logic
behind a method in the codeintel Context service, which will be the
thing that provides all access to SCIP data to Cody (in some manner,
depending on where the consumer is).

We will be able to split chunks more intelligently and also be able to
hydrate chunks with SCIP context if it's available at indexing time (in
the future). This PR lays scaffolding for that future.

## Test plan

CI; will have someone on the embeddings team confirm this is ok.