Cody JetBrains: add completions
Previously, the Cody JetBrains plugin only supported Chat. This PR adds
infrastructure to additionally provide inline completions using
JetBrains `Inlay` API. The current implementation is minimal and
provides a fairly bad user experience. The primary goal of this PR is to
add the infrastructure so that we can tweak and improve the UX with
smaller PRs.