Cody: Add E2E test setup based on Playwright (#52071)
This PR adds a Playwright test setup to Cody.

**How is this different from the existing integration test?**

Our existing integration tests already download and run VS Code and
allow us to create assertions inside the extension host runtime.
However, these tests are written on a lower level without easy ways to
interact with the UI.

Playwright is a very popular tool with exactly that in mind: interact
with the app as if you were a user. This way we can write very
high-level tests like this:

```ts
test('requires a valid auth token and allows logouts', async ({ page, sidebar }) => {
    await sidebar.getByRole('textbox', { name: 'Access Token (docs)' }).fill(VALID_TOKEN)
    await sidebar.getByRole('button', { name: 'Sign In' }).click()

    await expect(sidebar.getByText("Hello! I'm Cody.")).toBeVisible()
})
```

**Why Playwright?**

Playwright is the spiritual successor to Puppeteer and adds a lot of
nice APIs. It's also used to test various extensions from Microsoft and
VS Code itself.

There's also a lot of insane tooling that comes with using the
Playwright test runner that also works when it's attached to an Electron
app. Check out this example where I use the `--debug` mode to debug
selectors and even record new one based on interaction with the UI
(there are some not-so-smooth edges right now, e.g. like it opening an
empty Chromium tab in the example below 😅).


https://github.com/sourcegraph/sourcegraph/assets/458591/2e765ef0-e1a1-4e4c-94c0-2f269c42d52a

## Test plan

- This adds a new test which passes 

<!-- All pull requests REQUIRE a test plan:
https://docs.sourcegraph.com/dev/background-information/testing_principles
-->